+++ packages/docs/qui-docs/qui-docs.config.ts
@@ -1,3 +1,6 @@
+import {readFileSync} from "node:fs"
+import {resolve} from "node:path"
+
 import type {NavConfig, QuiDocsConfig} from "@qualcomm-ui/mdx-vite"
 
 const navConfig: NavConfig[] = [
@@ -63,6 +66,22 @@ export default {
         enabled: true,
         exclude: ["**/debug+/**", "*guide+/swagger*"],
       },
+      extraFiles: [
+        {
+          contents: readFileSync(
+            resolve(__dirname, "../../frameworks/react-mdx/CHANGELOG.md"),
+            "utf-8",
+          ),
+          id: "react-mdx-changelog",
+        },
+        {
+          contents: readFileSync(
+            resolve(__dirname, "../../common/mdx-vite/CHANGELOG.md"),
+            "utf-8",
+          ),
+          id: "mdx-vite-changelog",
+        },
+      ],
     },
   },
   navConfig,