+++ packages/common/tailwind-plugin/src/theme/theme-fonts.ts
@@ -2,112 +2,136 @@
 // SPDX-License-Identifier: BSD-3-Clause-Clear
 
 export interface FontData {
-  tailwind: string
+  tailwindClass: string
   variable: string
 }
 
 const display: FontData[] = [
-  {tailwind: "font-display-lg", variable: "--font-static-display-lg"},
-  {tailwind: "font-display-md", variable: "--font-static-display-md"},
+  {tailwindClass: "font-display-lg", variable: "--font-static-display-lg"},
+  {tailwindClass: "font-display-md", variable: "--font-static-display-md"},
 ]
 
 const dynamicDisplay: FontData[] = [
-  {tailwind: "font-dynamic-display-lg", variable: "--font-dynamic-display-lg"},
-  {tailwind: "font-dynamic-display-md", variable: "--font-dynamic-display-md"},
+  {
+    tailwindClass: "font-dynamic-display-lg",
+    variable: "--font-dynamic-display-lg",
+  },
+  {
+    tailwindClass: "font-dynamic-display-md",
+    variable: "--font-dynamic-display-md",
+  },
 ]
 
 const heading: FontData[] = [
-  {tailwind: "font-heading-xxxl", variable: "--font-static-heading-xxxl"},
-  {tailwind: "font-heading-xxl", variable: "--font-static-heading-xxl"},
-  {tailwind: "font-heading-xl", variable: "--font-static-heading-xl"},
-  {tailwind: "font-heading-lg", variable: "--font-static-heading-lg-default"},
-  {tailwind: "font-heading-md", variable: "--font-static-heading-md-default"},
-  {tailwind: "font-heading-sm", variable: "--font-static-heading-sm-default"},
-  {tailwind: "font-heading-xs", variable: "--font-static-heading-xs-default"},
+  {tailwindClass: "font-heading-xxxl", variable: "--font-static-heading-xxxl"},
+  {tailwindClass: "font-heading-xxl", variable: "--font-static-heading-xxl"},
+  {tailwindClass: "font-heading-xl", variable: "--font-static-heading-xl"},
   {
-    tailwind: "font-heading-xxs",
+    tailwindClass: "font-heading-lg",
+    variable: "--font-static-heading-lg-default",
+  },
+  {
+    tailwindClass: "font-heading-md",
+    variable: "--font-static-heading-md-default",
+  },
+  {
+    tailwindClass: "font-heading-sm",
+    variable: "--font-static-heading-sm-default",
+  },
+  {
+    tailwindClass: "font-heading-xs",
+    variable: "--font-static-heading-xs-default",
+  },
+  {
+    tailwindClass: "font-heading-xxs",
     variable: "--font-static-heading-xxs-default",
   },
 ]
 
 const headingBold: FontData[] = [
   {
-    tailwind: "font-heading-lg-bold",
+    tailwindClass: "font-heading-lg-bold",
     variable: "--font-static-heading-lg-bold",
   },
   {
-    tailwind: "font-heading-md-bold",
+    tailwindClass: "font-heading-md-bold",
     variable: "--font-static-heading-md-bold",
   },
   {
-    tailwind: "font-heading-sm-bold",
+    tailwindClass: "font-heading-sm-bold",
     variable: "--font-static-heading-sm-bold",
   },
   {
-    tailwind: "font-heading-xs-bold",
+    tailwindClass: "font-heading-xs-bold",
     variable: "--font-static-heading-xs-bold",
   },
   {
-    tailwind: "font-heading-xxs-bold",
+    tailwindClass: "font-heading-xxs-bold",
     variable: "--font-static-heading-xxs-bold",
   },
 ]
 
 const dynamicHeading: FontData[] = [
   {
-    tailwind: "font-dynamic-heading-xxxl",
+    tailwindClass: "font-dynamic-heading-xxxl",
     variable: "--font-dynamic-heading-xxxl",
   },
   {
-    tailwind: "font-dynamic-heading-xxl",
+    tailwindClass: "font-dynamic-heading-xxl",
     variable: "--font-dynamic-heading-xxl",
   },
-  {tailwind: "font-dynamic-heading-xl", variable: "--font-dynamic-heading-xl"},
   {
-    tailwind: "font-dynamic-heading-lg",
+    tailwindClass: "font-dynamic-heading-xl",
+    variable: "--font-dynamic-heading-xl",
+  },
+  {
+    tailwindClass: "font-dynamic-heading-lg",
     variable: "--font-dynamic-heading-lg-default",
   },
 ]
 
 const dynamicHeadingBold: FontData[] = [
   {
-    tailwind: "font-dynamic-heading-lg-bold",
+    tailwindClass: "font-dynamic-heading-lg-bold",
     variable: "--font-dynamic-heading-lg-bold",
   },
 ]
 
 const body: FontData[] = [
-  {tailwind: "font-body-xxl", variable: "--font-static-body-xxl-default"},
-  {tailwind: "font-body-xl", variable: "--font-static-body-xl-default"},
-  {tailwind: "font-body-lg", variable: "--font-static-body-lg-default"},
-  {tailwind: "font-body-md", variable: "--font-static-body-md-default"},
-  {tailwind: "font-body-sm", variable: "--font-static-body-sm-default"},
-  {tailwind: "font-body-xs", variable: "--font-static-body-xs-default"},
+  {tailwindClass: "font-body-xxl", variable: "--font-static-body-xxl-default"},
+  {tailwindClass: "font-body-xl", variable: "--font-static-body-xl-default"},
+  {tailwindClass: "font-body-lg", variable: "--font-static-body-lg-default"},
+  {tailwindClass: "font-body-md", variable: "--font-static-body-md-default"},
+  {tailwindClass: "font-body-sm", variable: "--font-static-body-sm-default"},
+  {tailwindClass: "font-body-xs", variable: "--font-static-body-xs-default"},
 ]
 
 const bodyBold: FontData[] = [
-  {tailwind: "font-body-xxl-bold", variable: "--font-static-body-xxl-bold"},
-  {tailwind: "font-body-xl-bold", variable: "--font-static-body-xl-bold"},
-  {tailwind: "font-body-lg-bold", variable: "--font-static-body-lg-bold"},
-  {tailwind: "font-body-md-bold", variable: "--font-static-body-md-bold"},
-  {tailwind: "font-body-sm-bold", variable: "--font-static-body-sm-bold"},
-  {tailwind: "font-body-xs-bold", variable: "--font-static-body-xs-bold"},
+  {
+    tailwindClass: "font-body-xxl-bold",
+    variable: "--font-static-body-xxl-bold",
+  },
+  {tailwindClass: "font-body-xl-bold", variable: "--font-static-body-xl-bold"},
+  {tailwindClass: "font-body-lg-bold", variable: "--font-static-body-lg-bold"},
+  {tailwindClass: "font-body-md-bold", variable: "--font-static-body-md-bold"},
+  {tailwindClass: "font-body-sm-bold", variable: "--font-static-body-sm-bold"},
+  {tailwindClass: "font-body-xs-bold", variable: "--font-static-body-xs-bold"},
 ]
 
 const code: FontData[] = [
-  {tailwind: "font-code-xl", variable: "--font-static-code-xl-default"},
-  {tailwind: "font-code-lg", variable: "--font-static-code-lg-default"},
-  {tailwind: "font-code-md", variable: "--font-static-code-md-default"},
-  {tailwind: "font-code-sm", variable: "--font-static-code-sm-default"},
-  {tailwind: "font-code-xs", variable: "--font-static-code-xs-default"},
+  {tailwindClass: "font-code-xl", variable: "--font-static-code-xl-default"},
+  {tailwindClass: "font-code-lg", variable: "--font-static-code-lg-default"},
+  {tailwindClass: "font-code-md", variable: "--font-static-code-md-default"},
+  {tailwindClass: "font-code-sm", variable: "--font-static-code-sm-default"},
+  {tailwindClass: "font-code-xs", variable: "--font-static-code-xs-default"},
 ]
 
 const codeBold: FontData[] = [
-  {tailwind: "font-code-xl-bold", variable: "--font-static-code-xl-bold"},
-  {tailwind: "font-code-lg-bold", variable: "--font-static-code-lg-bold"},
-  {tailwind: "font-code-md-bold", variable: "--font-static-code-md-bold"},
-  {tailwind: "font-code-sm-bold", variable: "--font-static-code-sm-bold"},
-  {tailwind: "font-code-xs-bold", variable: "--font-static-code-xs-bold"},
+  {tailwindClass: "font-code-xl-bold", variable: "--font-static-code-xl-bold"},
+  {tailwindClass: "font-code-lg-bold", variable: "--font-static-code-lg-bold"},
+  {tailwindClass: "font-code-md-bold", variable: "--font-static-code-md-bold"},
+  {tailwindClass: "font-code-sm-bold", variable: "--font-static-code-sm-bold"},
+  {tailwindClass: "font-code-xs-bold", variable: "--font-static-code-xs-bold"},
 ]
 
 export const themeFonts = {