+++ packages/frameworks/react-router-utils/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @qualcomm-ui/react-router-utils
+# @qualcomm-ui/react-router-utils Changelog
 
 ## 1.0.7 (2025/12/12)